Updated Get All Ratings

+++ demo/src/main/java/com/teleconsulting/demo/dto/DoctorRating.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @Builder
 public class DoctorRating {
     Long id;
-    int totalRating;
-    int appointmentCount;
+    float totalRating;
+    float appointmentCount;
 }
